+++ Sources/OpenAPIKit30/Security/SecurityScheme.swift
@@ -50,7 +50,7 @@ extension OpenAPI {
         }
 
         public enum SecurityType: Equatable {
-            case apiKey(name: String, location: SecuritySchemeLocation)
+            case apiKey(name: String, location: Location)
             case http(scheme: String, bearerFormat: String?)
             case oauth2(flows: OAuthFlows)
             case openIdConnect(openIdConnectUrl: URL)
