--- Sources/OpenAPIKitCore/JSONSchemaPermissions.swift
@@ -1,12 +1,0 @@
-//
-//  JSONSchemaPermissions.swift
-//  
-//
-//  Created by Mathew Polzin on 12/17/22.
-//
-
-public enum JSONSchemaPermissions: String, Codable {
-    case readOnly
-    case writeOnly
-    case readWrite
-}
