--- Sources/OpenAPIKitCore/ParameterSchemaContextStyle.swift
@@ -1,16 +1,0 @@
-//
-//  ParameterSchemaContextStyle.swift
-//  
-//
-//  Created by Mathew Polzin on 12/18/22.
-//
-
-public enum ParameterSchemaContextStyle: String, CaseIterable, Codable {
-    case form
-    case simple
-    case matrix
-    case label
-    case spaceDelimited
-    case pipeDelimited
-    case deepObject
-}
