+++ Sources/OpenAPIKitCore/Shared.swift
@@ -1,0 +1,11 @@
+//
+//  Shared.swift
+//  
+//
+//  Created by Mathew Polzin on 12/24/22.
+//
+
+/// A Core namespace for OpenAPI types that are shared by multiple OpenAPI standard versions.
+public enum Shared {
+
+}
