+++ Sources/OpenAPIKitCore/Shared/ContentType.swift
@@ -5,95 +5,97 @@
 //  Created by Mathew Polzin on 12/29/19.
 //
 
-/// The Content Type of an API request or response body.
-public struct ContentType: Codable, Equatable, Hashable, RawRepresentable, HasWarnings {
-    internal let underlyingType: Builtin
-    public let warnings: [Warning]
+extension Shared {
+    /// The Content Type of an API request or response body.
+    public struct ContentType: Codable, Equatable, Hashable, RawRepresentable, HasWarnings {
+        internal let underlyingType: Builtin
+        public let warnings: [Warning]
 
-    /// The type and subtype of the content type. This is everything except for
-    /// any parameters that are also attached.
-    ///
-    /// If you want the full content type string, use `rawValue`.
-    public var typeAndSubtype: String { underlyingType.rawValue }
+        /// The type and subtype of the content type. This is everything except for
+        /// any parameters that are also attached.
+        ///
+        /// If you want the full content type string, use `rawValue`.
+        public var typeAndSubtype: String { underlyingType.rawValue }
 
-    /// Key/Value pairs serialized as parameters for the content type.
-    ///
-    /// For exmaple, in "`text/plain; charset=UTF-8`" "charset" is
-    /// the name of a parameter with the value "UTF-8".
-    ///
-    /// OpenAPIKit will always encode these parameters with the keys in
-    /// alphabetical order.
-    public var parameters: [String: String]
+        /// Key/Value pairs serialized as parameters for the content type.
+        ///
+        /// For exmaple, in "`text/plain; charset=UTF-8`" "charset" is
+        /// the name of a parameter with the value "UTF-8".
+        ///
+        /// OpenAPIKit will always encode these parameters with the keys in
+        /// alphabetical order.
+        public var parameters: [String: String]
 
-    /// The full raw string value of the content type and any of its parameters.
-    ///
-    /// If you only want the type & subtype without any parameters, use
-    /// `typeAndSubtype`.
-    public var rawValue: String {
-        let rawParams = parameters
-            .sorted(by: { $0.0 < $1.0 })
-            .map { (name, value) in "; \(name)=\(value)" }
-            .joined()
+        /// The full raw string value of the content type and any of its parameters.
+        ///
+        /// If you only want the type & subtype without any parameters, use
+        /// `typeAndSubtype`.
+        public var rawValue: String {
+            let rawParams = parameters
+                .sorted(by: { $0.0 < $1.0 })
+                .map { (name, value) in "; \(name)=\(value)" }
+                .joined()
 
-        return typeAndSubtype + rawParams
-    }
-
-    public init?(rawValue: String) {
-        let parts = rawValue.split(separator: ";")
-        let type = parts.first.map(String.init) ?? rawValue
-
-        var warnings = [Warning]()
-        var params = [String:String]()
-        for part in parts.dropFirst() {
-            switch Self.parseParameter(part) {
-            case .success(let (name, value)):
-                params[name] = value
-            case .failure(let warning):
-                warnings.append(warning)
-            }
+            return typeAndSubtype + rawParams
         }
 
-        parameters = params
+        public init?(rawValue: String) {
+            let parts = rawValue.split(separator: ";")
+            let type = parts.first.map(String.init) ?? rawValue
 
-        if let underlying = Builtin.init(rawValue: type) {
-            underlyingType = underlying
-        } else {
-            underlyingType = .other(type)
-            warnings.append(
-                .message(
-                    "'\(rawValue)' could not be parsed as a Content Type. Content Types should have the format '<type>/<subtype>'"
+            var warnings = [Warning]()
+            var params = [String:String]()
+            for part in parts.dropFirst() {
+                switch Self.parseParameter(part) {
+                case .success(let (name, value)):
+                    params[name] = value
+                case .failure(let warning):
+                    warnings.append(warning)
+                }
+            }
+
+            parameters = params
+
+            if let underlying = Builtin.init(rawValue: type) {
+                underlyingType = underlying
+            } else {
+                underlyingType = .other(type)
+                warnings.append(
+                    .message(
+                        "'\(rawValue)' could not be parsed as a Content Type. Content Types should have the format '<type>/<subtype>'"
+                    )
+                )
+            }
+            self.warnings = warnings
+        }
+
+        private static func parseParameter(_ param: String.SubSequence) -> Result<(String, String), Warning> {
+            let parts = param.split(separator: "=")
+
+            guard parts.count == 2,
+                  let name = parts.first,
+                  let value = parts.last else {
+                return .failure(.message("Could not parse a Content Type parameter from '\(param)'"))
+            }
+
+            return .success(
+                (
+                    String(name.trimmingCharacters(in: .whitespaces)),
+                    String(value.trimmingCharacters(in: .whitespaces))
                 )
             )
         }
-        self.warnings = warnings
-    }
 
-    private static func parseParameter(_ param: String.SubSequence) -> Result<(String, String), Warning> {
-        let parts = param.split(separator: "=")
-
-        guard parts.count == 2,
-              let name = parts.first,
-              let value = parts.last else {
-                  return .failure(.message("Could not parse a Content Type parameter from '\(param)'"))
-              }
-
-        return .success(
-            (
-                String(name.trimmingCharacters(in: .whitespaces)),
-                String(value.trimmingCharacters(in: .whitespaces))
-            )
-        )
-    }
-
-    internal init(_ builtin: Builtin) {
-        underlyingType = builtin
-        parameters = [:]
-        warnings = []
+        internal init(_ builtin: Builtin) {
+            underlyingType = builtin
+            parameters = [:]
+            warnings = []
+        }
     }
 }
 
 // convenience constructors
-public extension ContentType {
+public extension Shared.ContentType {
     /// Bitmap image
     static let bmp: Self = .init(.bmp)
     static let css: Self = .init(.css)
@@ -155,7 +157,7 @@ public extension ContentType {
     static let any: Self = .init(.any)
 }
 
-extension ContentType {
+extension Shared.ContentType {
     // This internal representation makes it easier to ensure that the popular
     // builtin types supported are fully covered in their rawValue implementation.
     internal enum Builtin: Codable, Equatable, Hashable {
@@ -221,7 +223,7 @@ extension ContentType {
     }
 }
 
-extension ContentType.Builtin: RawRepresentable {
+extension Shared.ContentType.Builtin: RawRepresentable {
     public var rawValue: String {
         switch self {
         case .bmp: return "image/bmp"
@@ -314,4 +316,4 @@ extension ContentType.Builtin: RawRepresentable {
     }
 }
 
-extension ContentType: Validatable {}
+extension Shared.ContentType: Validatable {}
