+++ Tests/OpenAPIKitCoreTests/ContentTypeTests.swift
@@ -10,7 +10,7 @@ import XCTest
 
 final class ContentTypeTests: XCTestCase {
     func test_badContentType() {
-        let bad = ContentType(rawValue: "not a content type")
+        let bad = Shared.ContentType(rawValue: "not a content type")
         XCTAssertEqual(
             bad?.warnings[0].localizedDescription,
             "\'not a content type\' could not be parsed as a Content Type. Content Types should have the format \'<type>/<subtype>\'"
@@ -18,7 +18,7 @@ final class ContentTypeTests: XCTestCase {
     }
 
     func test_contentTypeStringReflexivity() {
-        let types: [ContentType] = [
+        let types: [Shared.ContentType] = [
             .bmp,
             .css,
             .csv,
@@ -57,24 +57,24 @@ final class ContentTypeTests: XCTestCase {
         ]
 
         for type in types {
-            XCTAssertEqual(type, ContentType(rawValue: type.rawValue))
+            XCTAssertEqual(type, Shared.ContentType(rawValue: type.rawValue))
         }
     }
 
     func test_goodParam() {
-        let type = ContentType.init(rawValue: "text/html; charset=utf8")
+        let type = Shared.ContentType.init(rawValue: "text/html; charset=utf8")
         XCTAssertEqual(type?.warnings.count, 0)
         XCTAssertEqual(type?.rawValue, "text/html; charset=utf8")
     }
 
     func test_multipleParams() {
-        let type = ContentType.init(rawValue: "my/type; some=thing; another=else")
+        let type = Shared.ContentType.init(rawValue: "my/type; some=thing; another=else")
         XCTAssertEqual(type?.warnings.count, 0)
         XCTAssert(type?.rawValue == "my/type; another=else; some=thing")
     }
 
     func test_oneBadParam() {
-        let type = ContentType.init(rawValue: "my/type; some: thing; another=else")
+        let type = Shared.ContentType.init(rawValue: "my/type; some: thing; another=else")
         XCTAssertEqual(type?.warnings.count, 1)
         XCTAssertEqual(type?.warnings.first?.localizedDescription, "Could not parse a Content Type parameter from \' some: thing\'")
         XCTAssert(type?.rawValue == "my/type; another=else")
